Modularize smoketest training into reusable pipeline
