+++ train_bestcase.py
@@ -1,0 +1,326 @@
+"""DVI best-case trainer with KL→RL schedule.
+
+Mapping of the original smoketest script to modular components:
+
+* Model assembly & dual-LoRA enablement → ``training.modeling``
+* KV hygiene helpers → ``training.kv``
+* Rollout & buffer operations → ``training.rollout``
+* Losses & schedules → ``training.objectives`` and ``training.schedule``
+* Evaluation (acceptance & CTARs) → ``evaluation.acceptance``
+* W&B + JSONL logging → ``training.logging``
+* General utilities (seed, prompts, env dump) → ``training.utils``
+"""
+import os
+os.environ.setdefault("PYTORCH_CUDA_ALLOC_CONF", "expandable_segments:True")
+
+import argparse
+import json
+import math
+import time
+from typing import List
+
+import torch
+
+from training.utils import (
+    set_seed,
+    ensure_dirs,
+    env_dump,
+    build_prompts_from_alpaca,
+    free_cuda,
+    _cuda_sync,
+)
+from training.modeling import prepare_dvi_trainable, build_optimizer
+from training.kv import estimate_kv_cache
+from training.rollout import rollout_collect, buf_debug
+from training.objectives import one_mixed_step
+from training.schedule import mix_schedule, phase_of_step
+from training.logging import init_wandb, wandb_watch_model, wandb_log, WANDB_DEFAULT_ENTITY, WANDB_DEFAULT_PROJECT
+from training.buffer import ReplayBuffer
+from evaluation.acceptance import eval_acceptance
+
+
+def train_bestcase_kl_rl(model, tok, prompts_train: List[str], prompts_eval: List[str],
+                        steps: int, rollout_len: int, batch_size: int,
+                        lr_exit: float, lr_lora: float, temperature: float,
+                        ce_weight: float, eval_every: int,
+                        warmup_kl: int, ramp_steps: int,
+                        kl_min: float, pg_max: float, ent_weight: float,
+                        outdir: str, accepted_only_flag: bool,
+                        debug_dump_every: int, debug_topk: int,
+                        max_fro: float, max_fro_ratio: float,
+                        quiet_eval: bool):
+    metrics_path = os.path.join(outdir, "logs", "train_metrics.jsonl")
+    samples_path = os.path.join(outdir, "logs", "rollout_samples.jsonl")
+
+    opt = build_optimizer(model, lr_exit=lr_exit, lr_lora=lr_lora, wd_exit=1e-2, wd_lora=0.0)
+    buf = ReplayBuffer(capacity=max(4096, batch_size * rollout_len * 8), device=torch.device("cpu"))
+
+    model.eval()
+    with torch.no_grad():
+        init_fro = float(torch.linalg.vector_norm(model.exit_proj.weight).item())
+        lm_norm = float(torch.linalg.vector_norm(model.lm_head.weight).item())
+    print(f"[sanity] head norms: ||lm||={lm_norm:.3f} ||exit||={init_fro:.3f}", flush=True)
+    wandb_log({"sanity/lm_head_norm": lm_norm, "sanity/exit_head_norm": init_fro}, step=0)
+
+    print("\n[e2e] measuring acceptance (pre-train)…", flush=True)
+    free_cuda("(pre-eval)")
+    acc0, ctar0 = eval_acceptance(
+        model, tok, prompts_eval, rollout_len, steps_per_prompt=1,
+        dump_debug=True, dump_path=samples_path, topk=debug_topk, quiet=quiet_eval
+    )
+    print(f"[e2e] PRE  | acc={acc0:.3f} | CTAR1={ctar0[1]:.3f} CTAR2={ctar0[2]:.3f} CTAR3={ctar0[3]:.3f} CTAR4={ctar0[4]:.3f}", flush=True)
+    wandb_log({"eval/pre/acc": acc0,
+               "eval/pre/ctar1": ctar0[1], "eval/pre/ctar2": ctar0[2],
+               "eval/pre/ctar3": ctar0[3], "eval/pre/ctar4": ctar0[4]}, step=0)
+
+    ptr = 0
+    tokens_total = 0
+    ema_tok_s = None
+
+    for g in range(steps):
+        model.eval()
+        p = prompts_train[ptr % len(prompts_train)]
+        ptr += 1
+        dbg_roll = [] if (debug_dump_every and (g % debug_dump_every == 0)) else None
+        _cuda_sync()
+        t_roll_s = time.perf_counter()
+        n_collected = rollout_collect(model, tok, p, buf, steps=rollout_len, debug_out=dbg_roll, topk=debug_topk)
+        _cuda_sync()
+        t_roll_e = time.perf_counter()
+        elapsed = max(1e-6, (t_roll_e - t_roll_s))
+        tok_s = float(n_collected) / elapsed
+        tokens_total += n_collected
+        ema_tok_s = tok_s if ema_tok_s is None else (0.9 * ema_tok_s + 0.1 * tok_s)
+
+        kv_bytes, kv_seq = estimate_kv_cache(model)
+        kv_mb = kv_bytes / (1024 ** 2)
+        if torch.cuda.is_available():
+            mem_alloc = torch.cuda.memory_allocated() / (1024 ** 2)
+            mem_reserved = torch.cuda.memory_reserved() / (1024 ** 2)
+            mem_max = torch.cuda.max_memory_allocated() / (1024 ** 2)
+        else:
+            mem_alloc = mem_reserved = mem_max = 0.0
+
+        if g % 10 == 0:
+            buf_debug(buf, k=16)
+
+        if len(buf) < batch_size:
+            print(f"[train] step {g:04d} | {phase_of_step(g, warmup_kl, ramp_steps):>10} | "
+                  f"buf={len(buf)}/{batch_size} | tok/s={tok_s:.2f} (EMA {ema_tok_s:.2f}) | "
+                  f"KV={kv_mb:.2f} MB (seq≈{kv_seq}) | waiting for batch…", flush=True)
+            wandb_log({"roll/tok_s": tok_s, "roll/tok_s_ema": ema_tok_s,
+                       "kv/bytes": kv_bytes, "kv/MB": kv_mb, "kv/seq_est": kv_seq,
+                       "cuda/mem_alloc_MB": mem_alloc, "cuda/mem_reserved_MB": mem_reserved,
+                       "cuda/mem_max_alloc_MB": mem_max,
+                       "buf/size": len(buf)}, step=g)
+            continue
+
+        lam_pg, lam_kl = mix_schedule(g, warmup_kl, ramp_steps, kl_min, pg_max)
+        phase = phase_of_step(g, warmup_kl, ramp_steps)
+        accepted_only = accepted_only_flag and (phase != "WARMUP(KL)") and (lam_pg > 0.0)
+        try:
+            batch = buf.sample(batch_size, accepted_only=accepted_only)
+        except ValueError:
+            if accepted_only:
+                print(f"[train][backoff] step {g:04d} | accepted-only requested but pool too small; sampling all tokens this step.")
+                batch = buf.sample(batch_size, accepted_only=False)
+            else:
+                print(f"[train] step {g:04d} | buffer insufficient (total={len(buf)}); collecting more…")
+                continue
+
+        model.train()
+        _cuda_sync()
+        t_trn_s = time.perf_counter()
+        out = one_mixed_step(
+            model, opt, batch,
+            temperature=temperature, clip=1.0,
+            ce_weight=ce_weight, lam_pg=lam_pg, lam_kl=lam_kl,
+            ent_weight=ent_weight,
+            init_fro=init_fro, max_fro=max_fro, max_fro_ratio=max_fro_ratio
+        )
+        _cuda_sync()
+        t_trn_e = time.perf_counter()
+        step_time = (t_trn_e - t_trn_s)
+
+        accept_roll = float(batch["reward"].float().mean().item())
+        buf_size = len(buf)
+        buf_acc = buf.accepted_count()
+        buf_rate = (buf_acc / buf_size) if buf_size else 0.0
+        with torch.no_grad():
+            exit_norm = float(torch.linalg.vector_norm(model.exit_proj.weight).item())
+
+        if not out["ok"] or not math.isfinite(out["loss"]) or not math.isfinite(out["kl"]):
+            print(f"[train][nan] step {g:04d} | {phase:>10} | "
+                  f"loss {out['loss']} | KL {out['kl']} | PG {out['pg']} | CE {out['ce']} | "
+                  f"grad {out['grad']} -> clearing opt state", flush=True)
+            opt.state.clear()
+            free_cuda("(nan-recovery)")
+            continue
+
+        std_s, std_t = out["std_s_t"]
+        print(
+            f"[train] step {g:04d} | {phase:>10} | "
+            f"loss {out['loss']:+7.4f} "
+            f"(PG {out['pg']:+7.4f}→{out['c_pg']:+7.4f} , KL {out['kl']:+7.4f}→{out['c_kl']:+7.4f} , "
+            f"CE {out['ce']:+6.4f}→{out['c_ce']:+6.4f} , ENT {out['ent']:+6.4f}→{out['c_ent']:+6.4f}) | "
+            f"grad {out['grad']:6.2f} | tok/s={tok_s:.2f} (EMA {ema_tok_s:.2f}) | step_time {step_time*1e3:6.1f} ms | "
+            f"π_s(v)≈{out['pi_v']:.4f} | std_s={std_s:.3f} std_t={std_t:.3f} | "
+            f"||exit||={exit_norm:.2f} | roll_acc≈{accept_roll:.4f} | "
+            f"buf_size={buf_size} acc_rate≈{buf_rate:.3f} | λ_pg={lam_pg:.3f} λ_kl={lam_kl:.3f} | accepted_only={accepted_only} | "
+            f"KV={kv_mb:.2f} MB (seq≈{kv_seq}) | CUDA alloc={mem_alloc:.1f} MB",
+            flush=True,
+        )
+
+        metrics = dict(
+            step=g, phase=phase,
+            loss=out["loss"], kl=out["kl"], pg=out["pg"], ce=out["ce"], ent=out["ent"],
+            c_pg=out["c_pg"], c_kl=out["c_kl"], c_ce=out["c_ce"], c_ent=out["c_ent"],
+            grad=float(out["grad"]), tok_s=tok_s, tok_s_ema=ema_tok_s, step_time_ms=step_time * 1e3,
+            pi_s_v=out["pi_v"], std_s=std_s, std_t=std_t,
+            exit_norm=exit_norm,
+            roll_acc=accept_roll, buf_size=buf_size, buf_acc_rate=buf_rate,
+            lam_pg=lam_pg, lam_kl=lam_kl, accepted_only=accepted_only,
+            kv_bytes=kv_bytes, kv_mb=kv_mb, kv_seq_est=kv_seq,
+            cuda_mem_alloc_mb=mem_alloc, cuda_mem_reserved_mb=mem_reserved, cuda_mem_max_alloc_mb=mem_max,
+            tokens_total=tokens_total,
+        )
+        with open(metrics_path, "a") as f:
+            f.write(json.dumps(metrics) + "\n")
+
+        wandb_log({
+            "loss/total": out["loss"], "loss/pg": out["pg"], "loss/kl": out["kl"],
+            "loss/ce": out["ce"], "loss/ent": out["ent"],
+            "contrib/pg": out["c_pg"], "contrib/kl": out["c_kl"],
+            "contrib/ce": out["c_ce"], "contrib/ent": out["c_ent"],
+            "grad/norm": float(out["grad"]),
+            "roll/tok_s": tok_s, "roll/tok_s_ema": ema_tok_s,
+            "time/step_ms": step_time * 1e3,
+            "policy/pi_s_v": out["pi_v"], "logits/std_s": std_s, "logits/std_t": std_t,
+            "head/exit_norm": exit_norm,
+            "roll/accept_rate_batch": accept_roll,
+            "buf/size": buf_size, "buf/accept_rate": buf_rate,
+            "schedule/lam_pg": lam_pg, "schedule/lam_kl": lam_kl,
+            "sampling/accepted_only": int(accepted_only),
+            "kv/bytes": kv_bytes, "kv/MB": kv_mb, "kv/seq_est": kv_seq,
+            "cuda/mem_alloc_MB": mem_alloc, "cuda/mem_reserved_MB": mem_reserved, "cuda/mem_max_alloc_MB": mem_max,
+            "tokens/total": tokens_total,
+        }, step=g)
+
+        if dbg_roll:
+            with open(samples_path, "a") as f:
+                for rec in dbg_roll:
+                    f.write(json.dumps({"step": g, **rec}) + "\n")
+
+        if (g + 1) % eval_every == 0:
+            model.eval()
+            free_cuda("(pre mid-train eval)")
+            print("[e2e] measuring acceptance (mid-train)…", flush=True)
+            acc_mid, ctar_mid = eval_acceptance(
+                model, tok, prompts_eval, rollout_len, steps_per_prompt=1,
+                dump_debug=True, dump_path=samples_path, topk=debug_topk, quiet=quiet_eval
+            )
+            print(f"[e2e] MID  | step {g+1:04d} | acc={acc_mid:.3f} | "
+                  f"CTAR1={ctar_mid[1]:.3f} CTAR2={ctar_mid[2]:.3f} CTAR3={ctar_mid[3]:.3f} CTAR4={ctar_mid[4]:.3f}", flush=True)
+            wandb_log({"eval/mid/acc": acc_mid,
+                       "eval/mid/ctar1": ctar_mid[1], "eval/mid/ctar2": ctar_mid[2],
+                       "eval/mid/ctar3": ctar_mid[3], "eval/mid/ctar4": ctar_mid[4]}, step=g + 1)
+
+    del opt
+    del buf
+    free_cuda("(pre post-train eval)")
+    model.eval()
+    print("\n[e2e] measuring acceptance (post-train)…", flush=True)
+    acc1, ctar1 = eval_acceptance(
+        model, tok, prompts_eval, rollout_len, steps_per_prompt=1,
+        dump_debug=True, dump_path=samples_path, topk=debug_topk, quiet=quiet_eval
+    )
+    print(f"[e2e] POST | acc={acc1:.3f} | Δ={acc1-acc0:+.3f} | "
+          f"CTAR1={ctar1[1]:.3f} CTAR2={ctar1[2]:.3f} CTAR3={ctar1[3]:.3f} CTAR4={ctar1[4]:.3f}", flush=True)
+    wandb_log({"eval/post/acc": acc1,
+               "eval/post/ctar1": ctar1[1], "eval/post/ctar2": ctar1[2],
+               "eval/post/ctar3": ctar1[3], "eval/post/ctar4": ctar1[4],
+               "eval/delta_acc": (acc1 - acc0)}, step=steps)
+
+
+def main():
+    ap = argparse.ArgumentParser()
+    ap.add_argument("--model-id", type=str, default="meta-llama/Llama-2-7b-hf")
+    ap.add_argument("--early-layer", type=int, default=4)
+    ap.add_argument("--train-prompts", type=int, default=512)
+    ap.add_argument("--eval-prompts", type=int, default=64)
+    ap.add_argument("--steps", type=int, default=300)
+    ap.add_argument("--rollout", type=int, default=8)
+    ap.add_argument("--batch-size", type=int, default=64)
+    ap.add_argument("--lr-exit", type=float, default=5e-4)
+    ap.add_argument("--lr-lora", type=float, default=5e-5)
+    ap.add_argument("--temperature", type=float, default=1.0)
+    ap.add_argument("--ce-weight", type=float, default=0.20)
+    ap.add_argument("--ent-weight", type=float, default=0.00)
+    ap.add_argument("--warmup-kl", type=int, default=40)
+    ap.add_argument("--ramp-steps", type=int, default=80)
+    ap.add_argument("--kl-min", type=float, default=0.05)
+    ap.add_argument("--pg-max", type=float, default=1.0)
+    ap.add_argument("--outdir", type=str, default="minbench_out")
+    ap.add_argument("--debug-dump-every", type=int, default=25)
+    ap.add_argument("--debug-topk", type=int, default=5)
+    ap.add_argument("--accepted-only", action="store_true")
+    ap.add_argument("--max-fro", type=float, default=0.0)
+    ap.add_argument("--max-fro-ratio", type=float, default=0.0)
+    ap.add_argument("--eval-every", type=int, default=50)
+    ap.add_argument("--seed", type=int, default=1234)
+    ap.add_argument("--no-wandb", action="store_true")
+    ap.add_argument("--wandb-entity", type=str, default=WANDB_DEFAULT_ENTITY)
+    ap.add_argument("--wandb-project", type=str, default=WANDB_DEFAULT_PROJECT)
+    ap.add_argument("--run-name", type=str, default=None)
+    ap.add_argument("--quiet-eval", action="store_true")
+    args = ap.parse_args()
+
+    ensure_dirs(args.outdir)
+    env_dump(args.outdir)
+    set_seed(args.seed)
+
+    from transformers import AutoTokenizer
+    tok = AutoTokenizer.from_pretrained(args.model_id, use_fast=True)
+    if tok.pad_token_id is None:
+        tok.pad_token = tok.eos_token
+
+    print("[e2e] loading prompts…", flush=True)
+    prompts_train = build_prompts_from_alpaca(args.train_prompts)
+    prompts_eval  = build_prompts_from_alpaca(args.eval_prompts)
+
+    print("[e2e] loading model…", flush=True)
+    model = prepare_dvi_trainable(args.model_id, args.early_layer, dtype=torch.float16)
+
+    with torch.no_grad():
+        lm_nrm   = model.lm_head.weight.detach().float().norm().item()
+        exit_nrm = model.exit_proj.weight.detach().float().norm().item()
+        diff_nrm = (model.lm_head.weight.detach().float() - model.exit_proj.weight.detach().float()).norm().item()
+    print("[sanity] head parity:", {"||lm||": lm_nrm, "||exit||": exit_nrm, "||lm-exit||": diff_nrm}, flush=True)
+
+    run = init_wandb(args)
+    wandb_log({"sanity/head_parity_lm": lm_nrm,
+               "sanity/head_parity_exit": exit_nrm,
+               "sanity/head_parity_diff": diff_nrm}, step=0)
+    wandb_watch_model(model, log_freq=25)
+
+    train_bestcase_kl_rl(
+        model, tok, prompts_train, prompts_eval,
+        steps=args.steps, rollout_len=args.rollout, batch_size=args.batch_size,
+        lr_exit=args.lr_exit, lr_lora=args.lr_lora, temperature=args.temperature,
+        ce_weight=args.ce_weight, eval_every=args.eval_every,
+        warmup_kl=args.warmup_kl, ramp_steps=args.ramp_steps,
+        kl_min=args.kl_min, pg_max=args.pg_max, ent_weight=args.ent_weight,
+        outdir=args.outdir, accepted_only_flag=args.accepted_only,
+        debug_dump_every=args.debug_dump_every, debug_topk=args.debug_topk,
+        max_fro=args.max_fro, max_fro_ratio=args.max_fro_ratio,
+        quiet_eval=args.quiet_eval
+    )
+
+    if run is not None:
+        try:
+            run.finish()
+        except Exception:
+            pass
+
+if __name__ == "__main__":
+    main()
